# backend/common/response/system/role.go
package system

import "time"

type RoleVo struct {
	RoleId     int64     `json:"roleId"`
	RoleName   string    `json:"roleName"`
	RoleKey    string    `json:"roleKey"`
	CreateTime time.Time `json:"createTime"`
	UpdateTime time.Time `json:"updateTime"`
}