# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用go语言搭建的管理后端",
        "title": "四零专属管理后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/system/auth/adminPwdLogin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authController"
                ],
                "summary": "运营端账号密码登录",
                "operationId": "adminPwdLogin",
                "parameters": [
                    {
                        "description": "密码登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.PwdLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-string"
                        }
                    }
                }
            }
        },
        "/system/permission/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissionController"
                ],
                "summary": "创建权限",
                "operationId": "createPermission",
                "parameters": [
                    {
                        "description": "创建权限参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.PermissionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/permission/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissionController"
                ],
                "summary": "删除权限",
                "operationId": "deletePermission",
                "parameters": [
                    {
                        "description": "删除权限参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.PermissionDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/permission/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissionController"
                ],
                "summary": "获取权限详情",
                "operationId": "getPermissionDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "权限ID",
                        "name": "permissionId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-system_PermissionVo"
                        }
                    }
                }
            }
        },
        "/system/permission/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissionController"
                ],
                "summary": "获取权限列表",
                "operationId": "getPermissionList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "权限名称",
                        "name": "permissionName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "权限标识",
                        "name": "perms",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-response_PageResult-system_PermissionVo"
                        }
                    }
                }
            }
        },
        "/system/permission/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permissionController"
                ],
                "summary": "更新权限",
                "operationId": "updatePermission",
                "parameters": [
                    {
                        "description": "更新权限参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.PermissionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/role/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roleController"
                ],
                "summary": "创建角色",
                "operationId": "createRole",
                "parameters": [
                    {
                        "description": "创建角色参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.RoleCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/role/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roleController"
                ],
                "summary": "删除角色",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "description": "删除角色参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.RoleDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/role/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roleController"
                ],
                "summary": "获取角色详情",
                "operationId": "getRoleDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-system_RoleVo"
                        }
                    }
                }
            }
        },
        "/system/role/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roleController"
                ],
                "summary": "获取角色列表",
                "operationId": "getRoleList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "roleName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色标识",
                        "name": "roleKey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-response_PageResult-system_RoleVo"
                        }
                    }
                }
            }
        },
        "/system/role/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roleController"
                ],
                "summary": "更新角色",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "description": "更新角色参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.RoleUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.ApiResult-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.ApiResult-response_PageResult-system_PermissionVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/response.PageResult-system_PermissionVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-response_PageResult-system_RoleVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/response.PageResult-system_RoleVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.ApiResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-system_PermissionVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/system.PermissionVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-system_RoleVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/system.RoleVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.PageResult-system_PermissionVo": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.PermissionVo"
                    }
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.PageResult-system_RoleVo": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.RoleVo"
                    }
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "system.PermissionCreateRequest": {
            "type": "object",
            "required": [
                "permissionName"
            ],
            "properties": {
                "permissionName": {
                    "type": "string"
                },
                "perms": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "system.PermissionDeleteRequest": {
            "type": "object",
            "required": [
                "permissionId"
            ],
            "properties": {
                "permissionId": {
                    "type": "integer"
                }
            }
        },
        "system.PermissionUpdateRequest": {
            "type": "object",
            "required": [
                "permissionId"
            ],
            "properties": {
                "permissionId": {
                    "type": "integer"
                },
                "permissionName": {
                    "type": "string"
                },
                "perms": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "system.PermissionVo": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "permissionId": {
                    "type": "integer"
                },
                "permissionName": {
                    "type": "string"
                },
                "perms": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "system.PwdLoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "system.RoleCreateRequest": {
            "type": "object",
            "required": [
                "roleKey",
                "roleName"
            ],
            "properties": {
                "roleKey": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
        "system.RoleDeleteRequest": {
            "type": "object",
            "required": [
                "roleId"
            ],
            "properties": {
                "roleId": {
                    "type": "integer"
                }
            }
        },
...
        "system.RoleUpdateRequest": {
            "type": "object",
            "required": [
                "roleId"
            ],
            "properties": {
                "roleId": {
                    "type": "integer"
                },
                "roleKey": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
        "system.RoleVo": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "roleId": {
                    "type": "integer"
                },
                "roleKey": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
    }
}